Refactoring network stack management into holder class
