+++ src/mqtt_client.rs
@@ -1,11 +1,12 @@
 use crate::{
     de::{deserialize::ReceivedPacket, PacketReader},
+    network_manager::InterfaceHolder,
     ser::serialize,
     session_state::SessionState,
-    Property, {debug, error, info, warn},
+    Error, Property, ProtocolError, {debug, error, info, warn},
 };
 
-use embedded_nal::{nb, IpAddr, SocketAddr, TcpClientStack};
+use embedded_nal::{IpAddr, SocketAddr, TcpClientStack};
 use embedded_time::{
     duration::{Extensions, Seconds},
     Clock,
@@ -29,48 +30,6 @@ pub enum QoS {
 
     /// A packet will be delivered exactly one time.
     ExactlyOne,
-}
-
-/// Possible errors encountered during an MQTT connection.
-#[derive(Debug, PartialEq)]
-pub enum Error<E> {
-    Network(E),
-    WriteFail,
-    NotReady,
-    Unsupported,
-    ProvidedClientIdTooLong,
-    Failed(u8),
-    Protocol(ProtocolError),
-    SessionReset,
-    Clock(embedded_time::clock::Error),
-}
-
-/// Errors that are specific to the MQTT protocol implementation.
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum ProtocolError {
-    Bounds,
-    DataSize,
-    Invalid,
-    Failed,
-    PacketSize,
-    MalformedPacket,
-    MalformedInteger,
-    UnknownProperty,
-    UnsupportedPacket,
-    BufferSize,
-    InvalidProperty,
-}
-
-impl<E> From<embedded_time::clock::Error> for Error<E> {
-    fn from(clock: embedded_time::clock::Error) -> Self {
-        Error::Clock(clock)
-    }
-}
-
-impl<E> From<ProtocolError> for Error<E> {
-    fn from(error: ProtocolError) -> Self {
-        Error::Protocol(error)
-    }
 }
 
 mod sm {
@@ -106,9 +65,8 @@ where
 
 /// A client for interacting with an MQTT Broker.
 pub struct MqttClient<N: TcpClientStack, C: Clock, const T: usize> {
-    network_stack: N,
+    pub(crate) network: InterfaceHolder<N>,
     clock: C,
-    socket: Option<N::TcpSocket>,
     session_state: SessionState<C>,
     connection_state: StateMachine<Context>,
 }
@@ -120,10 +78,7 @@ where
 {
     fn process(&mut self) -> Result<(), Error<N::Error>> {
         // Potentially update the state machine depending on the current socket connection status.
-        if !self
-            .socket_is_connected()
-            .map_err(|err| Error::Network(err))?
-        {
+        if !self.network.tcp_connected()? {
             self.connection_state.process_event(Events::Disconnect).ok();
         } else {
             self.connection_state.process_event(Events::Connect).ok();
@@ -132,18 +87,8 @@ where
         match self.connection_state.state() {
             // In the RESTART state, we need to reopen the TCP socket.
             &States::Restart => {
-                if let Some(socket) = self.socket.take() {
-                    self.network_stack
-                        .close(socket)
-                        .map_err(|err| Error::Network(err))?;
-                }
+                self.network.allocate_socket()?;
 
-                // Allocate a new socket to use and begin connecting it.
-                self.socket.replace(
-                    self.network_stack
-                        .socket()
-                        .map_err(|err| Error::Network(err))?,
-                );
                 self.connection_state
                     .process_event(Events::GotSocket)
                     .unwrap();
@@ -151,15 +96,8 @@ where
 
             // In the connect transport state, we need to connect our TCP socket to the broker.
             &States::ConnectTransport => {
-                self.network_stack
-                    .connect(
-                        self.socket.as_mut().unwrap(),
-                        SocketAddr::new(self.session_state.broker, 1883),
-                    )
-                    .map_err(|err| match err {
-                        nb::Error::WouldBlock => Error::WriteFail,
-                        nb::Error::Other(err) => Error::Network(err),
-                    })?;
+                self.network
+                    .connect(SocketAddr::new(self.session_state.broker, 1883))?;
             }
 
             // Next, connect to the broker via the MQTT protocol.
@@ -187,7 +125,8 @@ where
                 )?;
 
                 info!("Sending CONNECT");
-                self.write(packet)?;
+                self.network
+                    .write_packet(&mut self.session_state, &self.clock, packet)?;
 
                 self.connection_state
                     .process_event(Events::SentConnect)
@@ -200,36 +139,6 @@ where
         self.handle_timers()?;
 
         Ok(())
-    }
-
-    fn read(&mut self, mut buf: &mut [u8]) -> Result<usize, Error<N::Error>> {
-        // Atomically access the socket.
-        let socket = self.socket.as_mut().unwrap();
-        let result = self.network_stack.receive(socket, &mut buf);
-
-        result.or_else(|err| match err {
-            nb::Error::WouldBlock => Ok(0),
-            nb::Error::Other(err) => Err(Error::Network(err)),
-        })
-    }
-
-    fn write(&mut self, buf: &[u8]) -> Result<(), Error<N::Error>> {
-        let socket = self.socket.as_mut().unwrap();
-        self.network_stack
-            .send(socket, &buf)
-            .map_err(|err| match err {
-                nb::Error::WouldBlock => Error::WriteFail,
-                nb::Error::Other(err) => Error::Network(err),
-            })
-            .and_then(|written| {
-                if written != buf.len() {
-                    Err(Error::WriteFail)
-                } else {
-                    self.session_state
-                        .register_transmission(self.clock.try_now()?);
-                    Ok(())
-                }
-            })
     }
 
     /// Configure the MQTT keep-alive interval.
@@ -286,15 +195,17 @@ where
         let mut buffer: [u8; T] = [0; T];
         let packet = serialize::subscribe_message(&mut buffer, topic, packet_id, properties)?;
 
-        self.write(packet).and_then(|_| {
-            info!("Subscribing to `{}`: {}", topic, packet_id);
-            self.session_state
-                .pending_subscriptions
-                .push(packet_id)
-                .map_err(|_| Error::Unsupported)?;
-            self.session_state.increment_packet_identifier();
-            Ok(())
-        })
+        self.network
+            .write_packet(&mut self.session_state, &self.clock, packet)
+            .and_then(|_| {
+                info!("Subscribing to `{}`: {}", topic, packet_id);
+                self.session_state
+                    .pending_subscriptions
+                    .push(packet_id)
+                    .map_err(|_| Error::Unsupported)?;
+                self.session_state.increment_packet_identifier();
+                Ok(())
+            })
     }
 
     /// Determine if any subscriptions are waiting for completion.
@@ -310,10 +221,7 @@ where
     /// # Returns
     /// True if the client is connected to the broker.
     pub fn is_connected(&mut self) -> Result<bool, Error<N::Error>> {
-        Ok(self
-            .socket_is_connected()
-            .map_err(|err| Error::Network(err))?
-            && self.session_state.connected)
+        Ok(self.network.tcp_connected()? && self.session_state.connected)
     }
 
     /// Publish a message over MQTT.
@@ -352,16 +260,8 @@ where
 
         let mut buffer: [u8; T] = [0; T];
         let packet = serialize::publish_message(&mut buffer, topic, data, properties)?;
-        self.write(packet)
-    }
-
-    fn socket_is_connected(&mut self) -> Result<bool, N::Error> {
-        if self.socket.is_none() {
-            return Ok(false);
-        }
-
-        let socket = self.socket.as_ref().unwrap();
-        self.network_stack.is_connected(socket)
+        self.network
+            .write_packet(&mut self.session_state, &self.clock, packet)
     }
 
     fn reset(&mut self) {
@@ -497,7 +397,8 @@ where
                 self.session_state.ping_timeout.replace(now + PING_TIMEOUT);
 
                 let packet = serialize::ping_req_message(&mut buffer)?;
-                self.write(packet)?;
+                self.network
+                    .write_packet(&mut self.session_state, &self.clock, packet)?;
             }
         }
 
@@ -531,9 +432,8 @@ impl<N: TcpClientStack, C: Clock, const T: usize> Minimq<N, C, T> {
 
         let minimq = Minimq {
             client: MqttClient {
-                network_stack,
+                network: InterfaceHolder::new(network_stack),
                 clock,
-                socket: None,
                 session_state,
                 connection_state: StateMachine::new(Context),
             },
@@ -562,7 +462,7 @@ impl<N: TcpClientStack, C: Clock, const T: usize> Minimq<N, C, T> {
         }
 
         let mut buf: [u8; 1024] = [0; 1024];
-        let received = self.client.read(&mut buf)?;
+        let received = self.client.network.read(&mut buf)?;
         if received > 0 {
             debug!("Received {} bytes", received);
         }
